fix some size_t\int warnings.
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@245 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/include/switch_utils.h
@@ -70,7 +70,7 @@ SWITCH_DECLARE(switch_status) switch_socket_create_pollfd(switch_pollfd_t *poll,
 SWITCH_DECLARE(int) switch_socket_waitfor(switch_pollfd_t *poll, int ms);
 SWITCH_DECLARE(void) switch_swap_linear(int16_t *buf, int len);
 SWITCH_DECLARE(char *) switch_cut_path(char *in);
-SWITCH_DECLARE(int) switch_float_to_short(float *f, short *s, int len);
+SWITCH_DECLARE(size_t) switch_float_to_short(float *f, short *s, size_t len);
 SWITCH_DECLARE(int) switch_char_to_float(char *c, float *f, int len);
 SWITCH_DECLARE(int) switch_float_to_char(float *f, char *c, int len);
 SWITCH_DECLARE(int) switch_short_to_float(short *s, float *f, int len);
